+++ src/main/java/com/basketbandit/core/SystemVariables.java
@@ -1,6 +1,7 @@
 package com.basketbandit.core;
 
 import com.google.api.services.youtube.model.SearchResult;
+import pw.aru.api.nekos4j.Nekos4J;
 
 import java.util.HashMap;
 import java.util.List;
@@ -11,5 +12,7 @@ public class SystemVariables {
      * CommandSearch
      */
     public static HashMap<Long, List<SearchResult>> searchUsers = new HashMap<>();
+
+    public static Nekos4J nekoApi = new Nekos4J.Builder().build();
 
 }