+++ src/main/java/com/basketbandit/core/modules/nsfw/ModuleNSFW.java
@@ -4,6 +4,7 @@ import com.basketbandit.core.CommandExecutor;
 import com.basketbandit.core.modules.Command;
 import com.basketbandit.core.modules.Module;
 import com.basketbandit.core.modules.nsfw.commands.CommandEfukt;
+import com.basketbandit.core.modules.nsfw.commands.CommandNeko;
 import com.basketbandit.core.utils.Utils;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
@@ -11,7 +12,8 @@ public class ModuleNSFW extends Module {
 
     public ModuleNSFW(MessageReceivedEvent e, String[] command) {
         super("ModuleNSFW", "moduleNSFW", new Command[] {
-                new CommandEfukt()
+                new CommandEfukt(),
+                new CommandNeko()
         });
 
         if(e != null && command != null) {